LayoutTests/crypto/subtle/rsa-import-pkcs8-truncated-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test importing a truncated PKCS8 RSA-OAEP key that triggers bounds check failure");

// This test verifies that importing a malformed PKCS8 key with a truncated
// ASN.1 structure is rejected gracefully rather than crashing.
//
// The malformed key is crafted so that:
// 1. The first two size checks in importPkcs8() pass
// 2. But the final computed headerSize exceeds the buffer size
// 3. Without the fix, keyData.subspan(headerSize) crashes
//
// Key structure (22 bytes):
// - Offset 0: 0x30 (SequenceMark)
// - Offset 1: 0x00 (length byte, bytesUsedToEncodedLength returns 1)
// - Offset 2-20: 19 placeholder bytes (Version + RSAOIDHeader + OctetStringMark)
// - Offset 21: 0x82 (length byte, bytesUsedToEncodedLength returns 3)
//
// headerSize computation:
// 1. headerSize = 1
// 2. Check: 22 >= 2 (pass)
// 3. headerSize = 1 + 1 + 3 + 15 + 1 = 21
// 4. Check: 22 >= 22 (pass)
// 5. headerSize = 21 + 3 = 24
// 6. subspan(24) on 22-byte buffer -> CRASH (before fix)

var extractable = true;

// 22 bytes: 30 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 82
var truncatedPkcs8Key = hexStringToUint8Array("30000000000000000000000000000000000000000082");

// Should reject with DataError, not crash
shouldReject('crypto.subtle.importKey("pkcs8", truncatedPkcs8Key, {name: "RSA-OAEP", hash: "sha-256"}, extractable, ["decrypt"])');

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
